Move Gson and Jackson specific tests
Moved existing unit tests from `TestJson` to either `TestJavalinGson` or
`TestJavalinJackson` to group those object-mapper specific tests together.

Existing unit tests that were not specific to either object-mapper
remain in the `TestJson` class.
